Update ServoPresets.cs
Servo movement as a 0.01 programmed error. this causes to detect false presets when this error takes place. most of the times when this error happens, it's permanent/static and repetitive request will not cause anything to move.  
the 0.005 is considered for calculation error. summed with 0.01 will solve this issue.
i have reviewed the movement requests when the issue happens with adding debug.log to the code and certain spots and the logs kept saying requests are made a position while the servo stands at a distance 0.01 of this position. i looked the code up and saw that this error is programmed at the end of Interpolator.PrepareUpdate.